\name{BioFam}
\docType{data}
\alias{BioFam}
\title{Biographical survey of family life states}
\description{
Family life sequences from the Swiss Household Panel biographical survey. Life course sequences of family status for 2000 individuals, recorded from age 15 to 30.
The data were collected retrospectively in a survey carried out in 2002. The data set also contains two covariates, sex and religion, derived from the original data.
}
\usage{BioFam}
\format{Data frame of 2000 rows and 16 columns}
\source{Swiss Household Panel,  www.swisspanel.ch}
\references{
Muller, N. S., M. Studer, G. Ritschard (2007). Classification de parcours de vie a l'aide de l'optimal
matching. In XIVe Rencontre de la Socie te francophone de classification (SFC 2007), Paris, 5-7
septembre 2007, pp. 157-160. 

Gabadinho, A., Ritschard, G., Muller, N. S., Studer, M. (2011). Analyzing and Visualizing State Sequences in R with TraMineR. Journal of Statistical Software,
  40(4), 1-37. URL http://www.jstatsoft.org/v40/i04/.
}
\keyword{R, longitudinal data, BioFam}
