\name{gRapfa-package}
\alias{gRapfa-package}
\alias{gRapfa}
\docType{package}
\title{
Models for discrete longitudinal data using APFA.
}
\description{
The package supports the use of models for discrete longitudinal data using acyclic probabilistic finite automata (APFA).
}
\details{
\tabular{ll}{
Package: \tab gRapfa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-10\cr
License: GPL (>= 2)
Depends: R (>= 3.0.2), igraph

}
}
\references{
Ankinakatte, S. and Edwards, D. Modelling discrete longitudinal data using acyclic probabilistic finite automata. Submitted to "Computational Statistics and Data Analysis".

Edwards, D and Ankinakatte, S. Some context-specific graphical models for discrete longitudinal data. Submitted "Statistical Modelling: An International Journal". (Archive: http://arxiv.org/abs/1311.5066)
}
\author{
Smitha Ankinakatte <Smitha.AA@agrsci.dk> and David Edwards <David.Edwards@agrsci.dk>
Maintainer: Smitha Ankinakatte <Smitha.AA@agrsci.dk>
}
