\name{mildew.raw}
\docType{data}
\alias{mildew.raw}
\title{A cross between two isolates of the barley powdery Mildew fungus}
\description{
  The \code{mildew.raw} is stem from a cross between two isolates of the barley powdery Mildew fungus.
  For each of N = 70 offpring, p = 6 binary markers, each corresponding to a single locus, were
  recorded.}
\usage{mildew.raw}
\format{Data frame of 70 observations with 6 variables}
\source{Christiansen, S., Giese, H., 1990. Genetic analysis of the obligate parasitic barley powdery
mildew fungus based on RFLP and virulence loci. Theoretical and Applied Genetics 79 (5),
705-712.
}
\references{
Edwards, D., 1992. Linkage analysis using loglinear models. Computational Statistics & Data Analysis
13 (3), 281-290.
}
\keyword{R, longitudinal data, mildew}
