\name{Wheeze}
\docType{data}
\alias{Wheeze}
\title{Wheeze in Steubenville children}
\description{
  \code{wheeze} is a longitudinal data set of binary indicator of the presence of wheeze at ages 7,8,9
  and 10, among 537 Steubenville children.
}
\usage{Wheeze}
\format{Data frame of 537 observations with 4 variables}
\source{Ekholm A, Smith PWF, McDonald JW. Marginal regression analysis of a multivariate binary
response. Biometrika 1995; 82(4):847-854.
}
\references{
Ekholm A, McDonald JW, Smith PWF. Association models for a multivariate binary response. Biometrics
2000; 56:712-718.}
\keyword{R, longitudinal data, wheeze}
